refactor : fix segmentation method
